refactored banner in separate component
